Successfully receiving data from rabbitMQ and filtering
