+++ danger.txt
@@ -1,19 +1,41 @@
-kill
-terror
-bomb
-gun
-shoot
-shooting
-attack
-knife 
-stab
-rape
-victim
-cruel
-suffer
-damage
-dead
-death 
-injured
-injery
-catastrophe
+terror,
+attack,
+knife,
+stab,
+stabber,
+stabbing,
+rape,
+raping,
+rapist,
+victim,
+suffer,
+damage,
+dead,
+death,
+injured,
+injury,
+catastrophe,
+tsunami,
+gun,
+shoot,
+shooter,
+shooting,
+danger,
+a bomb,
+bomb threat,
+bombing,
+explosion,
+earthquake,
+assault,
+attack,
+warning,
+kill,
+killer,
+warning,
+lockdown,
+lock down,
+violence,
+tornado,
+riot,
+rape,
+threat
