+++ exceptions.csv
@@ -1,0 +1,5 @@
+#photoshoot
+photography
+xbox
+#NBA
+ps4
