+++ incidentGenerator.go
@@ -8,6 +8,7 @@ import (
 	"strconv"
 	"time"
 	"encoding/json"
+	"./models"
 )
 
 var SafeDictionary []string
@@ -15,65 +16,65 @@ var DangerDictionary []string
 var NumToGenerate int
 var PercentDangerous float64
 var JsonFileName string
-
-type Post struct {
-	Activity_url string
-	Author_favorites_count uint
-	Author_followers_count uint
-	Author_friends_count uint
-	Author_klout_score uint
-	Author_picture_url string
-	Author_posts_count uint
-	Author_profile_url string
-	Author_real_name string
-	Author_username string
-	Bio string
-	Cleaned_body_text string
-	Country string
-	Country_code string
-	Embedded_urls []string
-	Hashtags []string
-	Historical_search  bool
-	Is_reshare bool
-	Language string
-	Latitude float64
-	Like_count uint
-	Location_display_name string
-	Longitude float64
-	Mentions  []string
-	Meta_data []interface{}
-	Network string
-	Normalized_urls []string
-	Parent struct{
-		Parent_author   string
-		Parent_author_profile_url string
-		Parent_author_reach int
-		Parent_body_text string
-		Parent_created_at string
-		Parent_social_source_id string
-	}
-	Post_created_at string
-	Post_media  struct {
-		Media_url string
-		Url string
-		Display_url string
-		Media_type string
-		Video_url string
-	}
-	Raw_body_text string
-	Region string
-	Region_code string
-	Retweet_count uint
-	Social_monitor_sources []struct {
-		Company_uid string
-		Monitor_uid string
-		Keywords map[string]uint
-	}
-	Social_monitor_uids []string
-	Social_source_uid string
-	Source string
-	Topic_monitor_id string
-}
+//
+//type Post struct {
+//	Activity_url string
+//	Author_favorites_count uint
+//	Author_followers_count uint
+//	Author_friends_count uint
+//	Author_klout_score uint
+//	Author_picture_url string
+//	Author_posts_count uint
+//	Author_profile_url string
+//	Author_real_name string
+//	Author_username string
+//	Bio string
+//	Cleaned_body_text string
+//	Country string
+//	Country_code string
+//	Embedded_urls []string
+//	Hashtags []string
+//	Historical_search  bool
+//	Is_reshare bool
+//	Language string
+//	Latitude float64
+//	Like_count uint
+//	Location_display_name string
+//	Longitude float64
+//	Mentions  []string
+//	Meta_data []interface{}
+//	Network string
+//	Normalized_urls []string
+//	Parent struct{
+//		Parent_author   string
+//		Parent_author_profile_url string
+//		Parent_author_reach int
+//		Parent_body_text string
+//		Parent_created_at string
+//		Parent_social_source_id string
+//	}
+//	Post_created_at string
+//	Post_media  struct {
+//		Media_url string
+//		Url string
+//		Display_url string
+//		Media_type string
+//		Video_url string
+//	}
+//	Raw_body_text string
+//	Region string
+//	Region_code string
+//	Retweet_count uint
+//	Social_monitor_sources []struct {
+//		Company_uid string
+//		Monitor_uid string
+//		Keywords map[string]uint
+//	}
+//	Social_monitor_uids []string
+//	Social_source_uid string
+//	Source string
+//	Topic_monitor_id string
+//}
 
 func main() {
 	rand.Seed(time.Now().Unix())
@@ -88,10 +89,10 @@ func main() {
 	PercentDangerous = float64(dumb)
 	JsonFileName = string(os.Args[5])
 
-	var posts []Post
+	var posts []models.Post
 
 	for i := 0; i < NumToGenerate; i++ {
-		post := new(Post)
+		post := new(models.Post)
 		if float64(i) < float64(NumToGenerate) * PercentDangerous {
 			post.Raw_body_text = getDangerSentence()
 		} else {
