+++ main/testRabbit.go
@@ -1,0 +1,7 @@
+package main
+
+import "../Distributor"
+
+func main() {
+	Distributor.StartRabbit()
+}
